fix(minor): xmenu display
xmenu confirm space

back to lfcd

fix external only disable display

change class to floatingkitty

try fix waybar multi dispaly icon

focus float

fix swww wallpaper transition

!!remove kitty single instance

fix external only disable display

+++ 2gui-hypr/hypr/lib/ui_cmd.sh
@@ -32,9 +32,11 @@ _link_wallpaper() {
 	ln -sf "$PWD/$1" "$wallpaper_cache"
 }
 change_wallpaper() {
-  mv -f "$wallpaper_cache" "$wallpaper_cache-bak"
-  _link_wallpaper $1 || mv -f "$wallpaper_cache-bak" "$wallpaper_cache"
-	swww query || swww init && swww img $1 $SWWW_PARAMS
+	if [[ -f $1 ]]; then
+		mv -f "$wallpaper_cache" "$wallpaper_cache-bak"
+		_link_wallpaper $1 || mv -f "$wallpaper_cache-bak" "$wallpaper_cache"
+	fi
+	swww query || swww init && swww img $1 $SWWW_PARAMS 
 }
 
 change_layout() {
@@ -67,7 +69,7 @@ toggle_term() {
 	result=$(hyprctl -j clients | jq -c ".[] | select(.initialTitle == \"$tname\") | .pid")
 	focused=$(hyprctl -j clients | jq -c ".[] | select(.initialTitle == \"$tname\") | .focusHistoryID")
 	if [[ -z $result ]]; then
-		kitty -T $tname --class floating $2 &
+		kitty -T $tname --class floatingkitty $2 &
 		exit 0
 	else
 		hyprctl dispatch pin pid:$result
@@ -83,7 +85,7 @@ toggle_term_sp() {
 	tname="FTQCS$1"
 	result=$(hyprctl -j clients | jq -c ".[] | select(.initialTitle == \"$tname\") | .pid")
 	if [[ -z $result ]]; then
-		kitty -T $tname --class floating $2 &
+		kitty -T $tname --class floatingkitty $2 &
 		exit 0
 	else
 		hyprctl dispatch togglespecialworkspace $tname
@@ -93,6 +95,6 @@ open_term_sp() {
 	tname="FTQCS$1"
 	result=$(hyprctl -j clients | jq -c ".[] | select(.initialTitle == \"$tname\") | .pid")
 	if [[ -z $result ]]; then
-		kitty -T "FTQCS$1" --class floating $2 &
+		kitty -T "FTQCS$1" --class floatingkitty $2 &
 	fi
 }
